Upgrade Embree to 2.9

+++ libs/embree/common/simd/vboolf8_avx.h
@@ -59,7 +59,7 @@ namespace embree
       vh = _mm_lookupmask_ps[a >> 4];
 #endif
     }
-    
+
     __forceinline vboolf ( const vboolf4& a                  ) : v(_mm256_insertf128_ps(_mm256_castps128_ps256(a),a,1)) {}
     __forceinline vboolf ( const vboolf4& a, const vboolf4& b) : v(_mm256_insertf128_ps(_mm256_castps128_ps256(a),b,1)) {}
     __forceinline vboolf ( const __m128 a, const __m128 b) : vl(a), vh(b) {}
@@ -164,7 +164,7 @@ namespace embree
   __forceinline bool none      ( const vboolf8& valid, const vboolf8& b ) { return none(valid & b); }
 
   __forceinline unsigned int movemask( const vboolf8& a ) { return _mm256_movemask_ps(a); }
-  __forceinline size_t       popcnt  ( const vboolf8& a ) { return __popcnt(_mm256_movemask_ps(a)); }
+  __forceinline size_t       popcnt  ( const vboolf8& a ) { return __popcnt((size_t)_mm256_movemask_ps(a)); }
 
   ////////////////////////////////////////////////////////////////////////////////
   /// Get/Set Functions
